+++ src/AElf.Kernel.SmartContract.Orleans.Core/AElf.Kernel.SmartContract.Orleans.Core.csproj
@@ -3,8 +3,9 @@
     <PropertyGroup>
         <TargetFramework>net6.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
-        <RootNamespace>AElf.Kernel.SmartContract.Grains</RootNamespace>
-        <PackageId>AElf.Kernel.SmartContract.Grains</PackageId>
+        <RootNamespace>AElf.Kernel.SmartContract.Orleans</RootNamespace>
+        <PackageId>AElf.Kernel.SmartContract.Orleans.Core</PackageId>
+        <AssemblyName>AElf.Kernel.SmartContract.Orleans.Core</AssemblyName>
     </PropertyGroup>
 
     <ItemGroup>
