+++ src/AElf.Kernel.SmartContract.Orleans.Core/CoreSmartContractOrleansAElfModule.cs
@@ -1,6 +1,7 @@
 using AElf.Contracts.Genesis;
 using AElf.Kernel.SmartContract.Application;
 using AElf.Kernel.SmartContract.Infrastructure;
+using AElf.Kernel.SmartContract.Orleans.Strategy;
 using AElf.Runtime.CSharp;
 using Microsoft.Extensions.DependencyInjection;
 using Orleans.Runtime;
@@ -8,10 +9,10 @@ using Orleans.Runtime.Placement;
 using Volo.Abp;
 using Volo.Abp.Modularity;
 
-namespace AElf.Kernel.SmartContract.Grains;
+namespace AElf.Kernel.SmartContract.Orleans;
 
 [DependsOn(typeof(CSharpRuntimeAElfModule), typeof(SmartContractAElfModule))]
-public class GrainExecutionAElfModule : AbpModule
+public class CoreSmartContractOrleansAElfModule : AbpModule
 {
     public override void ConfigureServices(ServiceConfigurationContext context)
     {
