+++ src/AElf.Silo.Launcher/AElfSiloLauncherModule.cs
@@ -10,7 +10,7 @@ using AElf.Kernel.SmartContract;
 using AElf.Kernel.SmartContract.ExecutionPluginForCallThreshold;
 using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee;
 using AElf.Kernel.SmartContract.ExecutionPluginForResourceFee;
-using AElf.Kernel.SmartContract.Grains;
+using AElf.Kernel.SmartContract.Orleans;
 using AElf.Kernel.Token;
 using AElf.RuntimeSetup;
 using Microsoft.Extensions.Configuration;
@@ -33,7 +33,7 @@ namespace AElf.Silo.Launcher;
     typeof(RuntimeSetupAElfModule),
     typeof(GovernmentSystemAElfModule),
     typeof(EconomicSystemAElfModule),
-    typeof(GrainExecutionAElfModule),
+    typeof(CoreSmartContractOrleansAElfModule),
     typeof(CoreConsensusAElfModule),
     typeof(AbpAspNetCoreModule)
 )]
